Rename package to apriltag_ros - Relates to https://github.com/AprilRobotics/apriltag_ros/issues/45
