+++ apriltag_ros/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(apriltags2_ros)
+project(apriltag_ros)
 
 find_package(catkin REQUIRED COMPONENTS
   geometry_msgs
