+++ apriltag_ros/config/settings.yaml
@@ -1,6 +1,6 @@
-# AprilTags 2 code parameters
-# Find descriptions in apriltag2/include/apriltag.h:struct apriltag_detector
-#                      apriltag2/include/apriltag.h:struct apriltag_family
+# AprilTag 3 code parameters
+# Find descriptions in apriltag/include/apriltag.h:struct apriltag_detector
+#                      apriltag/include/apriltag.h:struct apriltag_family
 tag_family:        'tag36h11' # options: tag36h11, tag25h9, tag16h5
 tag_threads:       2          # default: 2
 tag_decimate:      1.0        # default: 1.0
