+++ apriltag_ros/include/apriltag_ros/single_image_detector.h
@@ -41,13 +41,13 @@
  * Originator:        Danylo Malyuta, JPL
  ******************************************************************************/
 
-#ifndef APRILTAGS2_ROS_SINGLE_IMAGE_DETECTOR_H
-#define APRILTAGS2_ROS_SINGLE_IMAGE_DETECTOR_H
+#ifndef APRILTAG_ROS_SINGLE_IMAGE_DETECTOR_H
+#define APRILTAG_ROS_SINGLE_IMAGE_DETECTOR_H
 
-#include "apriltags2_ros/common_functions.h"
-#include <apriltags2_ros/AnalyzeSingleImage.h>
+#include "apriltag_ros/common_functions.h"
+#include <apriltag_ros/AnalyzeSingleImage.h>
 
-namespace apriltags2_ros
+namespace apriltag_ros
 {
 
 class SingleImageDetector
@@ -62,10 +62,10 @@ class SingleImageDetector
   SingleImageDetector(ros::NodeHandle& nh, ros::NodeHandle& pnh);
 
   // The function which provides the single image analysis service
-  bool analyzeImage(apriltags2_ros::AnalyzeSingleImage::Request& request,
-                     apriltags2_ros::AnalyzeSingleImage::Response& response);
+  bool analyzeImage(apriltag_ros::AnalyzeSingleImage::Request& request,
+                     apriltag_ros::AnalyzeSingleImage::Response& response);
 };
 
-} // namespace apriltags2_ros
+} // namespace apriltag_ros
 
-#endif // APRILTAGS2_ROS_SINGLE_IMAGE_DETECTOR_H
+#endif // APRILTAG_ROS_SINGLE_IMAGE_DETECTOR_H
