+++ apriltag_ros/scripts/analyze_image.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+#
+# Detect AprilTag in a single image. Usage:
+#
+# ./analyze_image.sh <image_load_path> <image_save_path>
+#
+
+roslaunch apriltag_ros single_image_client.launch image_load_path:="$1" image_save_path:="$2"
