+++ apriltag_ros/src/apriltag_ros_continuous_node.cpp
@@ -29,17 +29,22 @@
  * Technology.
  */
 
-#include "apriltags2_ros/common_functions.h"
-#include "apriltags2_ros/single_image_detector.h"
+#include <ros/ros.h>
+
+#include <nodelet/loader.h>
 
 int main(int argc, char **argv)
 {
-  ros::init(argc, argv, "apriltags2_ros_single_image_server");
+  ros::init(argc, argv, "apriltag_ros");
 
-  ros::NodeHandle nh;
-  ros::NodeHandle pnh("~");
+  nodelet::Loader nodelet;
+  nodelet::M_string remap(ros::names::getRemappings());
+  nodelet::V_string nargv;
 
-  apriltags2_ros::SingleImageDetector continuous_tag_detector(nh, pnh);
-  
+  nodelet.load(ros::this_node::getName(),
+              "apriltag_ros/ContinuousDetector",
+              remap, nargv);
+
   ros::spin();
+  return 0;
 }
