--- apriltags2_ros/launch/single_image_server.launch
@@ -1,12 +1,0 @@
-<launch>
-
-  <arg name="launch_prefix" default="" /> <!--set to value="gdbserver localhost:10000" for remote debugging-->
-  <arg name="node_namespace" default="apriltags2_ros_single_image_server_node" />
-
-  <!-- Set parameters -->
-  <rosparam command="load" file="$(find apriltags2_ros)/config/settings.yaml" ns="$(arg node_namespace)" />
-  <rosparam command="load" file="$(find apriltags2_ros)/config/tags.yaml" ns="$(arg node_namespace)" />
-
-  <node pkg="apriltags2_ros" type="apriltags2_ros_single_image_server_node" name="$(arg node_namespace)" clear_params="true" output="screen" launch-prefix="$(arg launch_prefix)" />
-
-</launch>
