--- apriltags2_ros/nodelet_plugins.xml
@@ -1,9 +1,0 @@
-<library path="lib/libapriltags2_ros_continuous_detector">
-    <class name="apriltags2_ros/ContinuousDetector"
-        type="apriltags2_ros::ContinuousDetector"
-		base_class_type="nodelet::Nodelet">
-		<description>
-          Continuous AprilTags 2 detector
-		</description>
-	</class>
-</library>
